# Resolve high `CacheMemoryContext` usage for `ANALYZE` on large partition table.