@JunfengYang 2020-07-21
I test this patch with below partition table, it contains 3999 leaf tables.
```sql
create table foopart (
    i int4, c1 int, c2 int, c3 int, c4 int, c5 int,
    c6 int, c7 int, c8 int, c9 int, c10 int,
    c11 int, c12 int, c13 int, c14 int, c15 int,
    c16 int, c17 int, c18 int, c19 int, c20 int,
    c21 int, c22 int, c23 int, c24 int, c25 int,
    c26 int, c27 int, c28 int, c29 int, c30 int,
    c31 int, c32 int, c33 int, c34 int, c35 int,
    c36 int, c37 int, c38 int, c39 int, c40 int,
    c41 int, c42 int, c43 int, c44 int, c45 int,
    c46 int, c47 int, c48 int, c49 int, c50 int,
    c51 int, c52 int, c53 int, c54 int, c55 int,
    c56 int, c57 int, c58 int, c59 int, c60 int,
    c61 int, c62 int, c63 int, c64 int, c65 int,
    c66 int, c67 int, c68 int, c69 int, c70 int,
    c71 int, c72 int, c73 int, c74 int, c75 int,
    c76 int, c77 int, c78 int, c79 int, c80 int,
    c81 int, c82 int, c83 int, c84 int, c85 int,
    c86 int, c87 int, c88 int, c89 int, c90 int
)
distributed by (i) partition by range (i) (start (1) end (4000) every (1));
```
After insert for 2000(foopart_1_prt_1 to foopart_1_prt_2000)leaf tables using below SQL with `aotostats` enabled.
```sql
INSERT INTO foopart_1_prt_(num)
    select (num), i % 1001, i % 1002, i % 1003, i % 1004,  i % 1005,
           i % 1006, i % 1007, i % 1008, i % 1009,  i % 1010,
           i % 1011, i % 1012, i % 1013, i % 1014,  i % 1015,
           i % 1016, i % 1017, i % 1018, i % 1019,  i % 1020,
           i % 1021, i % 1022, i % 1023, i % 1024,  i % 1025,
           i % 1026, i % 1027, i % 1028, i % 1029,  i % 1030,
           i % 1031, i % 1032, i % 1033, i % 1034,  i % 1035,
           i % 1036, i % 1037, i % 1038, i % 1039,  i % 1040,
           i % 1041, i % 1042, i % 1043, i % 1044,  i % 1045,
           i % 1046, i % 1047, i % 1048, i % 1049,  i % 1050,
           i % 1051, i % 1052, i % 1053, i % 1054,  i % 1055,
           i % 1056, i % 1057, i % 1058, i % 1059,  i % 1060,
           i % 1061, i % 1062, i % 1063, i % 1064,  i % 1065,
           i % 1066, i % 1067, i % 1068, i % 1069,  i % 1070,
           i % 1071, i % 1072, i % 1073, i % 1074,  i % 1075,
           i % 1076, i % 1077, i % 1078, i % 1079,  i % 1080,
           i % 1081, i % 1082, i % 1083, i % 1084,  i % 1085,
           i % 1086, i % 1087, i % 1088, i % 1089,  i % 1090
    from generate_series(1, 1000) as i;
```

Then insert a single leaf table(foopart_1_prt_2001) in a new session.
With the patch, the CacheMemoryContext used size is: 33,205,320.
Without the patch, the size is: 345,552,472.
Almost 10 times larger.


Insert 5 leaf tables (foopart_1_prt_3001 to foopart_1_prt_3005) in the same session:
With the patch:
```
INSERT 0 1000
Time: 4726.216 ms

INSERT 0 1000
Time: 4656.542 ms

INSERT 0 1000
Time: 4634.682 ms

INSERT 0 1000
Time: 4692.932 ms

INSERT 0 1000
Time: 4726.068 ms
```

Without the patch:
```
INSERT 0 1000
Time: 8172.572 ms

INSERT 0 1000
Time: 639.381 ms

INSERT 0 1000
Time: 607.146 ms

INSERT 0 1000
Time: 602.592 ms

INSERT 0 1000
Time: 660.714 ms
```
This is because we don't rely on cache anymore, so for leaf tables in the same root partition table, the `leaf_parts_analyzed `
execution time will be same for each insert.

Let's discuss which we should choose, better performance for specific cases, or save memory for analysis on large root partition.