@chrishajas 2020-07-21
Is the gpcopy/gprestore performance impact ok here? Not caching merge_leaf_stats seems fine since for gprestore/gpcopy the actual merge step is only run once on the final partition. However the leaf_parts_analyzed logic is run for every partition, so that is a fixed cost that will be hit on every partition regardless of data size. It should be ok/not noticeable unless there is little data and many partitions, but it might scale poorly as the number of partitions increase (4000 partitions * 4s = 4h, maybe 1h if we run it in parallel).

If users have their own scripts that analyze after ETL (rather than using analyzedb), they might also run into this issue.

Possibly relevant use cases to consider:

analyzedb - sets optimizer_analyze_enable_merge_of_leaf_stats=false and  runs analyze rootpartition at the end, so this change won't affect that utility.

gpcopy/gprestore - will run copies (in parallel) on empty partitions. So it will incur the leaf_parts_analyzed overhead for each partition, but the merge_leaf_stats overhead only once. Might be an issue for many partitions

ETL workloads where new data inserted into a few of partitions - seems ok since only copying into a few partitions.

custom scripts that manually run analyze on individual partition table leafs - is this a legitimate concern?

I'm wondering what other cases we should be aware of, since we've seen a few different use cases with analyze and don't want to impact the performance too much.