@JunfengYang 2020-07-22
@chrishajas 
Right now, gpcopy and gprestore already disable `autostats` and do ANALYZE on the root partition when copy all it's leaf tables.

I'm also worried if users use their own ETL tools will encounter the performance issue. But they may encounter the memory issue currently.
`leaf_parts_analyzed` will break the loop when it finds a leaf table that doesn't have stats. So if only a few leaf tables are analyzed,
the performance is ok. But ETL will face this issue eventually.

Maybe we can document the best practice for ETL usage. Another choice is we can leave this PR and wait for more cases from user side.
We already notice that some users encounter the memory issue when using the previous version of gprestore and gpcopy.