@zhangh43 2020-07-22
My thought:
1. For performance issue, since not all the partition table are deep(large number of leafs) and wide(large number of column). Could we leverage the two level access to catalog table. E.g. For the number of leaf partitions(columns) is less than a threshold, we still use syscache and for large partitions we use index scan?
2. Do you have test result on PG? How does upstream handle this deep+wide problem.