@JunfengYang 2020-07-22
@zhangh43 PG always samples data for ANALYZE, it does not rely on system catalog cache heavily.
We do the merge stats for the root partition table, which read leaf stats from leaf tables.