@JunfengYang 2020-07-22
Maybe we can add a guc to control the limit for using cache or not.
The limit should consider the (table_column_number * number_of_leaf_table).
Any thoughts?