@JunfengYang 2020-07-22
But if the (table_column_number * number_of_leaf_table) is small, current patch will also have good performance.